requirements.txt
cryptography>=42.0.0
protobuf>=4.25.0
grpcio-tools>=1.60.0
mss>=9.0.0
Pillow>=10.0.0
pyyaml>=6.0
pyinstaller>=6.0.0
